Features/eslint select verification code

+++ backend/src/application/auth/usecases/deleteAccount.ts
@@ -1,31 +1,55 @@
-import { DataHasher } from "@app/interfaces/security/DataHasher";
-import { BadBodyException } from "@cloneoverflow/common";
-import { UserRepository } from "@core/domain/repositories/user/UserRepository";
-import { AuthServiceInput } from "../dto/AuthServiceInput";
-import { AuthServiceOutput } from "../dto/AuthServiceOutput";
-import { IDeleteAccountUseCase } from "../types/usecases";
+import { DataHasher } from '@application/interfaces/security/DataHasher';
+import { BadBodyException, LoginException, RetriesExpiredException, VerificationCodeType } from '@cloneoverflow/common';
+import { CacheRepository } from '@core/domain/repositories/cache/CacheRepository';
+import { UserRepository } from '@core/domain/repositories/user/UserRepository';
+import { VerificationCodePayload } from '../data/VerificationCodePayload';
+import { AuthServiceInput } from '../dto/AuthServiceInput';
+import { AuthServiceOutput } from '../dto/AuthServiceOutput';
+import { IDeleteAccountUseCase } from '../types/usecases';
 
 export class DeleteAccountUseCase implements IDeleteAccountUseCase {
   constructor (
     private userRepository: UserRepository,
+    private cacheRepository: CacheRepository,
     private dataHasher: DataHasher,
   ) {}
 
-  async execute({ userId, creds }: AuthServiceInput.DeleteAccount): Promise<AuthServiceOutput.DeleteAccount> {
+  async execute (
+    { executorId, code, email, password }: AuthServiceInput.DeleteAccount,
+  ): Promise<AuthServiceOutput.DeleteAccount> {
     const confirmUser = await this.userRepository.findWithCreds({
-      where: { id: userId },
+      where: { id: executorId },
     });
   
     if (!confirmUser) {
-      throw new BadBodyException("Invalid user id");
+      throw new BadBodyException('Invalid user id');
     }
-    if (confirmUser.creds.email !== creds.email){
-      throw new BadBodyException("Invalid email or password");
+
+    const resolveCode = await this.cacheRepository.getObject<VerificationCodePayload>(
+      `user:${VerificationCodeType.DeletePassword}:${executorId}`,
+    );
+    
+    if (!resolveCode) {
+      throw new BadBodyException('User does not have verification code');
     }
-    if (!await this.dataHasher.compareHash(creds.password, confirmUser.creds.password)){
-      throw new BadBodyException("Invalid email or password");
-    }
+
+    if (resolveCode.retries <= 0) {
+      await this.cacheRepository.delete(`user:${VerificationCodeType.ForgotPassword}:${executorId}`);
   
+      throw new RetriesExpiredException();
+    }
+
+    if (await this.dataHasher.compareHash(code, resolveCode.code)) {
+      throw new BadBodyException('Code does not match');
+    }
+
+    if (
+      confirmUser.creds.email !== email || 
+      !await this.dataHasher.compareHash(password, confirmUser.creds.password)
+    ) {
+      throw new LoginException();
+    }
+
     await this.userRepository.delete({ 
       user: confirmUser.user,
     });
